pyproject.toml
[project]
name = "poke-mcp"
version = "0.1.0"
description = "Model Context Protocol server for Pokemon VGC team analysis"
authors = [{ name = "Sam", email = "sam@example.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "fastmcp>=0.1.1",
    "google-generativeai>=0.7.2",
    "chromadb>=0.5.5",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.5.7",
]

[project.scripts]
poke-mcp = "poke_mcp.server:run"

[build-system]
requires = ["setuptools>=65.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["poke_mcp"]
exclude = ["data"]
